% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ArchR_palettes_copy}
\alias{ArchR_palettes_copy}
\title{List of color palettes that can be used in plots}
\format{
An object of class \code{list} of length 30.
}
\usage{
ArchR_palettes_copy
}
\description{
A collection of some original and some borrowed color palettes to provide appealing color aesthetics for plots in ArchR
}
\keyword{datasets}
