% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{construct_pal}
\alias{construct_pal}
\title{A palette generator}
\usage{
construct_pal(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are: lots.}

\item{n}{Number of colors desired.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes for the Buenrostrolab and from ArchR and others!
}
\examples{
construct_pal("Cavalcanti")
construct_pal("Cavalcanti", 3)
construct_pal("Cavalcanti", c(1,3))

}
\keyword{colors}
