% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{jdb_palettes_copy}
\alias{jdb_palettes_copy}
\title{List of color palettes that can be used in plots from buencolors}
\format{
An object of class \code{list} of length 106.
}
\usage{
jdb_palettes_copy
}
\description{
A collection of some original and some borrowed color palettes to provide appealing color aesthetics for plots
}
\keyword{datasets}
